+++ package.json
@@ -34,7 +34,7 @@
     "build:renderer": "webpack --config webpack.config.renderer.babel.js",
     "build:main": "webpack --config webpack.config.main.babel.js",
     "test": "npm run lint",
-    "lint": "eslint src --ext .jsx",
+    "lint": "eslint src --ext .js,.jsx",
     "package": "npm run package:mac",
     "package:mac": "NODE_ENV=production npm run build; electron-packager . --overwrite --asar --app-version=`node -pe 'require(\"./package.json\").version'` --app-copyright=\"© `node -pe '(new Date).getFullYear()'` fiahfy\" --out release --ignore release --platform=darwin --arch=x64 --icon=app/assets/img/hosty.icns",
     "package:win": "NODE_ENV=production npm run build; electron-packager . --overwrite --asar --app-version=`node -pe 'require(\"./package.json\").version'` --app-copyright=\"© `node -pe '(new Date).getFullYear()'` fiahfy\" --out release --ignore release --platform=win32 --arch=x64 --icon=app/assets/img/hosty.ico --version-string.FileDescription=\"Electron application\" --version-string.OriginalFilename=\"Hosty.exe\" --version-string.ProductName=\"Hosty\"",
@@ -55,6 +55,7 @@
     "electron-packager": "8.5.1",
     "electron-rebuild": "^1.5.11",
     "file-loader": "^0.11.2",
+    "history": "^4.6.2",
     "is-electron-renderer": "^2.0.1",
     "material-ui": "^0.18.3",
     "prop-types": "^15.5.8",
