+++ src/main/app.js
@@ -1,7 +1,7 @@
 import { app } from 'electron';
 import fs from 'fs';
 import path from 'path';
-import AppWindow from './app-window';
+import Window from './window';
 import * as HostsFileManager from '../renderer/utils/hosts-file-manager';
 
 const PATH_WINDOW_OPTIONS = path.join(app.getPath('userData'), 'window.json');
@@ -17,7 +17,7 @@ export default class App {
 
     const options = App.loadWindowOptions();
 
-    this.window = new AppWindow(this);
+    this.window = new Window(this);
     this.window.open(options);
   }
   removeWindow() {
